Delete fixed

+++ src/routes/user-routes.js
@@ -843,12 +843,24 @@ router.delete('/:userId/groups/:groupId', async (req, res, next) => {
       })
     )
     await Member.deleteOne({ user_id, group_id })
-    let pathsList = await Path.find({group_id: group_id, car_owner_id: user_id})
-    pathsList.forEach(path =>{
-      Waypoint.deleteMany({path_id: path.path_id})
+
+    const userCars = await Car.find({ owner_id: user_id }).then((userCars) => {
+      userCars.forEach(car => {
+        const pathList = await Path.find({ car_id: car.car_id, group_id: group_id }).then((pathList) => {
+          pathList.forEach((path) => {
+            Waypoint.deleteMany({ path_id: path.path_id })
+          })
+        })
+        await Path.deleteMany({ group_id: group_id, car_id: car.car_id })
+      })
+    });
+
+    const groupPathList = await Path.find({ group_id: group_id }).then((groupPathList) => {
+      groupPathList.forEach(path => {
+        await Waypoint.deleteMany({ path_id: path.path_id, passenger_id: user_id })
+      })
     })
-    await Path.deleteMany({group_id: group_id, car_owner_id: user_id})
-    await Waypoint.deleteMany({passenger_id: user_id})
+
     res.status(200).send('User left group')
   } catch (error) {
     next(error)
@@ -1481,7 +1493,7 @@ router.post(
     }
 
     try {
-      await Car.create({...car, car_id: objectid(), owner_id})
+      await Car.create({ ...car, car_id: objectid(), owner_id })
       res.status(200).send('Car created')
     } catch (error) {
       next(error)
